refactor(ui): split layout state and fix policy compliance
