+++ modules/qt/include/ui/MainWindowLayoutState.hpp
@@ -1,0 +1,10 @@
+/*
+ * @file modules/qt/include/ui/MainWindowLayoutState.hpp
+ * Mirror header for MainWindow layout state helpers.
+ */
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUTSTATE_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUTSTATE_HPP_
+
+#include "ui/MainWindow.hpp"
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_MAINWINDOWLAYOUTSTATE_HPP_
