+++ modules/qt/include/ui/UiEnums.hpp
@@ -2,7 +2,8 @@
  * @file modules/qt/include/ui/UiEnums.hpp
  * Centralized UI enums and converters for readability and consistency.
  */
-#pragma once
+#ifndef GRAVITY_MODULES_QT_INCLUDE_UI_UIENUMS_HPP_
+#define GRAVITY_MODULES_QT_INCLUDE_UI_UIENUMS_HPP_
 
 #include <string>
 
@@ -12,72 +13,14 @@ enum class Solver { PairwiseCuda, OctreeGpu, OctreeCpu };
 enum class Integrator { Euler, Rk4 };
 enum class PerformanceProfile { Interactive, Balanced, Quality, Custom };
 
-// Converters: enum -> string
-static inline std::string to_string(Solver s)
-{
-    switch (s) {
-    case Solver::PairwiseCuda:
-        return "pairwise_cuda";
-    case Solver::OctreeGpu:
-        return "octree_gpu";
-    case Solver::OctreeCpu:
-        return "octree_cpu";
-    }
-    return {};
-}
+std::string to_string(Solver s);
+std::string to_string(Integrator i);
+std::string to_string(PerformanceProfile p);
 
-static inline std::string to_string(Integrator i)
-{
-    switch (i) {
-    case Integrator::Euler:
-        return "euler";
-    case Integrator::Rk4:
-        return "rk4";
-    }
-    return {};
-}
-
-static inline std::string to_string(PerformanceProfile p)
-{
-    switch (p) {
-    case PerformanceProfile::Interactive:
-        return "interactive";
-    case PerformanceProfile::Balanced:
-        return "balanced";
-    case PerformanceProfile::Quality:
-        return "quality";
-    case PerformanceProfile::Custom:
-        return "custom";
-    }
-    return {};
-}
-
-// Converters: string -> enum (returns default on unknown)
-static inline Solver solver_from_string(const std::string& s)
-{
-    if (s == "pairwise_cuda")
-        return Solver::PairwiseCuda;
-    if (s == "octree_gpu")
-        return Solver::OctreeGpu;
-    return Solver::OctreeCpu;
-}
-
-static inline Integrator integrator_from_string(const std::string& s)
-{
-    if (s == "rk4")
-        return Integrator::Rk4;
-    return Integrator::Euler;
-}
-
-static inline PerformanceProfile performance_from_string(const std::string& s)
-{
-    if (s == "balanced")
-        return PerformanceProfile::Balanced;
-    if (s == "quality")
-        return PerformanceProfile::Quality;
-    if (s == "custom")
-        return PerformanceProfile::Custom;
-    return PerformanceProfile::Interactive;
-}
+Solver solver_from_string(const std::string& s);
+Integrator integrator_from_string(const std::string& s);
+PerformanceProfile performance_from_string(const std::string& s);
 
 } // namespace grav_qt
+
+#endif // GRAVITY_MODULES_QT_INCLUDE_UI_UIENUMS_HPP_
